# Conflicts:
#	src/main/java/com/aptner/v3/global/error/ErrorCode.java

+++ src/main/java/com/aptner/v3/user/repository/UserDetailsRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserDetailsRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserDetailsRepository extends JpaRepository<UserEntity, Long> {
     Optional<UserEntity> findByUsername(String username);
 }
